+++ server/src/route.js
@@ -15,29 +15,7 @@ router.post('/graphql', graphqlHTTP({
     graphiql: false
 }))
 
-const client = require('../config/elasticsearch-config')
-const ProductSearchDTO = require('./product-search-dto')
-
-router.get('/search/:query', async (req, res, next) => {
-    try {
-        const result = await client.search({
-            index: 'product',
-            body: {
-                query: {
-                    match: {
-                        "name.nori": req.params.query
-                    }
-                }
-            }
-        })
-        const dtoList = result.body.hits.hits.map((hit, idx) => {
-            return new ProductSearchDTO(hit._source)
-        })
-
-        res.json(dtoList)
-    } catch (error) {
-        next(error)
-    }
-})
+const searchInIndex = require('./elasticsearch-controller')
+router.get('/search/:query', searchInIndex)
 
 module.exports = router
